when S3 class

Creates a when object.

Usage

when(
  name = NULL,
  type = NULL,
  locale = NULL,
  start = lubridate::today(),
  end = lubridate::today(),
  values = NULL,
  ...
)

Arguments

name

A string, table name.

type

A string, type of calendar (NULL, 'iso', 'epi' or 'time').

locale

A locale, to use for day and month names.

start

A string, start of the period to be included in the dimension.

end

A string, end of the period to be included in the dimension.

values

A vector of string.

...

Rest of boolean configuration parameters.

Value

A when object.

Details

Using the parameters of this function we can configure practically all the elements of the dimension. Alternatively, we can use the configuration functions to define the available options.

We discuss the parameters in each of the specific configuration functions.

See also

Examples

td_1 <- when()

td_2 <- when(type = 'time')